run.py
import sys
import os
import uvicorn
from dotenv import load_dotenv

# Charger les variables d'environnement depuis .env
load_dotenv()

# Ajouter le répertoire parent au PYTHONPATH
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

if __name__ == "__main__":
    print("Démarrage de l'API Fenn...")
    print("L'API sera disponible à l'adresse: http://localhost:8000")
    print("Documentation Swagger: http://localhost:8000/docs")
    uvicorn.run("backend.app.main:app", host="0.0.0.0", port=8000, reload=True) 