requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
firebase-admin==6.4.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
email-validator==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.3.2
transformers==4.37.2
python-slugify==8.0.1
pillow==10.2.0
python-jwt==4.1.0
cryptography==42.0.2 